issues v0.0.1 Issues.CLI

Handles the command line parsing and dispatch to various functions that end up generating the table of last n issues in a GitHub project

Functions

convert_to_list_of_maps(list)

decode_response(arg)

main(argv)

parse_args(argv)

argv can be -h or —help, which returns :help. Otherwise it is a github user name, project name, and (optional) the number of entries to format. Return a tuple of { user, project, count }, or :help if help was given.

process(arg1)

sort_into_ascending_order(issue_list)